requirements.txt
fastapi
pydantic
requests
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
